git-svn-id: http://svn.sagaengine.com/sagaengine/trunk@528 65b84acb-c606-4e19-8676-f724b2686c9c

+++ core/src/io/parse/component/StatComponentParserModule.cpp
@@ -81,7 +81,7 @@ namespace se_core {
 					Property* p = 0;
 					switch(propertyType) {
 					case Property::PT_SHORT:
-						p = new Property(name.get(), in.readShort());
+						p = new Property(name.get(), (short)in.readShort());
 						break;
 					case Property::PT_INT:
 						p = new Property(name.get(), in.readInt());
